# .docs/migration/openapi.yaml
...
openapi: 3.1.0
info:
  title: YACC REST API (canonical)
  version: 1.0.0-mig-003-canonical
  description: |
    Frozen canonical REST contract for the SPEC-002 Java/Spring backend and the adapted
    frontend. Derived from the MIG-001 behavior capture of the Node/TypeScript backend
    (routing-controllers) and canonicalized in MIG-003. The Java target re-expresses this
    contract via springdoc from `@RestController`s (ADR-023); the frontend consumes
    generated TS types from this file (MIG-034/MIG-041).

    Contract conventions (frozen):
    - No global route prefix; `/api` is declared at controller level (`@JsonController('/api/...')`
      at baseline, `@RequestMapping('/api/...')` in Java).
      Exception: health endpoints declare `/health` (no `/api`).
    - List endpoints return the BaseListResponse envelope `{data, page, limit, total}`
      (page is 1-indexed in responses; internally controllers pass 0-indexed page).
    - Simple resource envelopes: `{data: <entity>}` for single results.
    - Error shape: `{error: string}` (routing-controllers default error handler);
      IRC integration endpoints use `{code: string, message: string}`.
    - Auth: `Authorization: Bearer <accessToken>`; the POC's BETTER_AUTH_SECRET
      crypto-coupling is unwound in Java per ADR-025 (invisible to the wire contract).
    - Roles on the wire are lowercase: super_admin | admin | manager | user.
      User ids are UUID strings.
...
  x-capture-task: MIG-001
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/forgot-password:
    post:
      tags: [auth]
      operationId: authForgotPassword
      summary: Request password reset (public; always 200 to prevent enumeration; rate limited)
      security: []
      x-baseline-source: packages/backend/src/controllers/auth.controller.ts:44 @98db628
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ForgotPasswordBody'}
      responses:
        '200':
          description: Constant anti-enumeration message
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MessageResponse'}
        '429':
          description: Rate limit exceeded (passwordResetRateLimiter)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/reset-password:
    post:
      tags: [auth]
      operationId: authResetPassword
      summary: Reset password with token (public; rate limited)
      security: []
      x-baseline-source: packages/backend/src/controllers/auth.controller.ts:100 @98db628
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordBody'}
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SuccessMessageResponse'}
        '400':
          description: Invalid or expired token (generic, anti-enumeration)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '429':
          description: Rate limit exceeded (passwordResetRateLimiter)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/verify-email:
    post:
      tags: [auth]
      operationId: authVerifyEmail
      summary: Confirm an email address with the emailed token (public; MIG-031 deviation)
      description: |
        MIG-031 email-verification confirmation. The challenge is issued on
        self-registration (sign-up) as a single-use, expiring 256-bit email
        token; this endpoint consumes it (replay-safe by deletion), flips
        `email_verified`, and audit-logs the event. Baseline lived inside the
        BetterAuth wildcard (REST-AUTH-004); recorded per the MIG-031
        contract-change note in this file's header.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VerifyEmailBody'}
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SuccessMessageResponse'}
        '400':
          description: Invalid or expired token (generic, anti-enumeration)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/sign-in/email:
    post:
      tags: [auth]
      operationId: authSignInEmail
      summary: Email/password sign-in (public; rate limited)
      description: |
        Canonical sign-in. Baseline delegated to BetterAuth; the Java target implements
        this endpoint via Spring Security (ADR-025). Response DTO canonicalized in MIG-003
        (ledger row REST-AUTH-003): `{user, accessToken, refreshToken}`; token
        claims/lifecycle are ADR-025/MIG-033 implementation detail.
      security: []
      x-baseline-source: packages/backend/src/controllers/auth.controller.ts:136 @98db628
      x-canonicalization: MIG-003 (REST-AUTH-003; rate limit frozen 5 req/15 min)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string}
      responses:
        '200':
          description: Signed in — user object plus stateless JWT access/refresh tokens
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '429':
          description: Rate limit exceeded (5 requests / 15 minutes)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/sign-up/email:
    post:
      tags: [auth]
      operationId: authSignUpEmail
      summary: Self-registration (public; creates USER role only)
      description: |
        Canonical self-registration (SPEC-002 FR-04/AC-11). Replaces the baseline
        BetterAuth wildcard sub-path `sign-up/email` (ledger row REST-AUTH-004) with an
        explicit endpoint per ADR-025. The request accepts NO role field; the created user
        is always `role: user`. Privileged user/role management stays super_admin-only
        (`POST /api/users`).
      security: []
      x-baseline-source: BetterAuth wildcard handler (controllers/auth.controller.ts:154 @98db628, sub-path sign-up/email)
      x-canonicalization: MIG-003 (REST-AUTH-004 wire sub-path mapping)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: {type: string, format: email}
                password: {type: string}
                name: {type: string}
      responses:
        '201':
          description: Registered — user created with role `user` plus tokens
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        '400':
          description: Validation error (e.g. email already registered — generic error)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/sign-out:
    post:
      tags: [auth]
      operationId: authSignOut
      summary: Sign out (authenticated)
      description: |
        Canonical sign-out. Replaces the baseline BetterAuth wildcard sub-path `sign-out`
        (ledger row REST-AUTH-004). Invalidates the presented grant per ADR-025 token
        lifecycle; with stateless JWT access tokens this is an acknowledgment + refresh
        revocation.
      security:
        - bearerAuth: []
      x-baseline-source: BetterAuth wildcard handler (controllers/auth.controller.ts:154 @98db628, sub-path sign-out)
      x-canonicalization: MIG-003 (REST-AUTH-004 wire sub-path mapping)
      responses:
        '200':
          description: Signed out
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSignOutResponse'}
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/get-session:
    get:
      tags: [auth]
      operationId: authGetSession
      summary: Current session user (authenticated)
      description: |
        Canonical session lookup. Replaces the baseline BetterAuth wildcard sub-path
        `get-session` (ledger row REST-AUTH-004). Returns the user resolved from the
        Bearer access token; there is no server-side session object in the canonical
        stateless model (ADR-025).
      security:
        - bearerAuth: []
      x-baseline-source: BetterAuth wildcard handler (controllers/auth.controller.ts:154 @98db628, sub-path get-session)
      x-canonicalization: MIG-003 (REST-AUTH-004 wire sub-path mapping)
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionGetResponse'}
        '401':
          description: Missing, invalid, expired token, or inactive/suspended account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/refresh-token:
    post:
      tags: [auth]
      operationId: authRefreshToken
      summary: Refresh access token (public; refresh-grant rotation)
      description: |
        Canonical token refresh. Replaces the baseline BetterAuth wildcard sub-path
        `refresh-token` (ledger row REST-AUTH-004). Implements the ADR-025 refresh grant
        (stateless JWT access + refresh default) with rotation.
      security: []
      x-baseline-source: BetterAuth wildcard handler (controllers/auth.controller.ts:154 @98db628, sub-path refresh-token)
      x-canonicalization: MIG-003 (REST-AUTH-004 wire sub-path mapping)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: {type: string}
      responses:
        '200':
          description: Tokens rotated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthRefreshResponse'}
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/auth/change-password:
    post:
      tags: [auth]
      operationId: authChangePassword
      summary: Replace the credential (authenticated; MIG-030 contract-change finding)
      description: |
        Authenticated password change that clears the ADR-025 forced-password-change
        state of the bootstrap/recovery identity (MIG-030). The current credential is
        verified; every refresh grant of the identity is revoked on success. MIG-031's
        email-token reset/verification flows are separate and unchanged.
      security:
        - bearerAuth: []
      x-mig-030-contract-change: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: {type: string}
                newPassword: {type: string, minLength: 8}
      responses:
        '200':
          description: Credential replaced; forced-change state cleared
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSignOutResponse'}
        '400':
          description: Validation error (e.g. password below minimum length)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401':
          description: Missing/invalid token or wrong current password
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

  # =====================================================================
  # UsersController — @JsonController('/api/users')
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Baseline JWT verified against BETTER_AUTH_SECRET (ADR-025 unwinds this coupling;
        wire contract `Authorization: Bearer` is preserved by the Java target).
  schemas:
...
    ErrorResponse:
      type: object
      description: Default routing-controllers error shape
      required: [error]
      properties:
        error: {type: string}
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: {type: string}
    SuccessMessageResponse:
      type: object
      required: [success, message]
      properties:
        success: {type: boolean}
        message: {type: string}

    # ---------- Auth (canonicalized in MIG-003; ledger rows REST-AUTH-001..004) ----------
    AuthSessionResponse:
      type: object
      description: "Sign-in / self-registration response. `user.role` is always lowercase; sign-up always creates `role: user`. Token claims/lifecycle are ADR-025/MIG-033 implementation detail. `mustChangePassword` is a MIG-030 contract-change finding (ADR-025 forced-password-change state)."
      required: [user, accessToken, refreshToken, mustChangePassword]
      properties:
        user: {$ref: '#/components/schemas/UserResponse'}
        accessToken: {type: string, description: Stateless JWT access token (ADR-025 default)}
        refreshToken: {type: string, description: Refresh token (rotated on refresh-grant use)}
        mustChangePassword: {type: boolean, description: True while the bootstrap/recovery identity must replace its credential (MIG-030)}
    AuthSignOutResponse:
      type: object
      required: [success]
      properties:
        success: {type: boolean, enum: [true]}
    AuthSessionGetResponse:
      type: object
      description: "Session lookup — the user resolved from the Bearer access token (no server-side session object; ADR-025 stateless default). `mustChangePassword` is a MIG-030 contract-change finding."
      required: [user, mustChangePassword]
      properties:
        user: {$ref: '#/components/schemas/UserResponse'}
        mustChangePassword: {type: boolean, description: True while the bootstrap/recovery identity must replace its credential (MIG-030)}
    AuthRefreshResponse:
      type: object
      description: Rotated token pair.
      required: [accessToken, refreshToken]
      properties:
        accessToken: {type: string}
        refreshToken: {type: string}

    # ---------- Health ----------
...
    ForgotPasswordBody:
      type: object
      required: [email]
      properties:
        email: {type: string, format: email}
    ResetPasswordBody:
      type: object
      required: [token, password]
      properties:
        token: {type: string, description: 64-char hex reset token}
    VerifyEmailBody:
      type: object
      required: [token]
      properties:
        token: {type: string, description: 64-char hex verification token (MIG-031 deviation)}
        password: {type: string}

    # ---------- Users ----------
    UserRole:
      type: string
      enum: [super_admin, admin, manager, user]
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
    UserResponse:
      type: object
      description: Baseline shape packages/backend/src/types/users.types.ts UserResponse
      required: [id, email, name, role, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        name: {type: string}
        role: {$ref: '#/components/schemas/UserRole'}
        status: {$ref: '#/components/schemas/UserStatus'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        deletedAt: {type: [string, 'null'], format: date-time}
...